+++ services/api/src/app.module.ts
@@ -4,10 +4,11 @@ import { HealthController } from "./http/health.controller.js";
 import { FamilyController } from "./http/family.controller.js";
 import { TaskController } from "./http/task.controller.js";
 import { RewardController } from "./http/reward.controller.js";
+import { AuthController } from "./http/auth.controller.js";
 import { ErrorFilter } from "./http/error.filter.js";
 
 @Module({
-  controllers: [HealthController, FamilyController, TaskController, RewardController],
+  controllers: [HealthController, AuthController, FamilyController, TaskController, RewardController],
   providers: [{ provide: APP_FILTER, useClass: ErrorFilter }],
 })
 export class AppModule {}